+++ fiddles/angular2/fiddle-0019-ZingChart/config.js
@@ -1,0 +1,20 @@
+
+System.config({
+    //use typescript for compilation
+    transpiler: 'typescript',
+    //typescript compiler options
+    typescriptOptions: {
+        emitDecoratorMetadata: true
+    },
+    //map tells the System loader where to look for things
+    map: {
+        app: "./src"
+    },
+    //packages defines our app package
+    packages: {
+        app: {
+            main: './boot.ts',
+            defaultExtension: 'ts'
+        }
+    }
+});