+++ fiddles/angular2/fiddle-0019-ZingChart/src/meta.ts
@@ -1,0 +1,9 @@
+
+var metadata = {
+    fiddleHeader: 'Angular 2 - AgGrid / ZingChart',
+    urls: {
+        data: 'data.json',
+        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0019-ZingChart'
+    },
+    consoleTag: 'H O U S E ~ f i d d l e s'
+};